Split each unit test into files
