+++ tests/Unit/Dashboard/User_Interface/Abstract_Site_Kit_Widget_Permanent_Dismissal_Route_Test.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface;
+
+use Mockery;
+use WP_Error;
+use Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Widget_Permanent_Dismissal_Route;
+use Yoast\WP\SEO\Tests\Unit\Dashboard\Infrastructure\Repositories\Permanently_Dismissed_Site_Kit_Widget_Repository_Fake;
+use Yoast\WP\SEO\Tests\Unit\TestCase;
+
+/**
+ * Abstract class for the Permanently Dismissed Site Kit Widget Repository tests.
+ *
+ * @group site_kit_widget_permanent_dismissal
+ *
+ * @phpcs:disable Yoast.NamingConventions.ObjectNameDepth.MaxExceeded
+ */
+abstract class Abstract_Site_Kit_Widget_Permanent_Dismissal_Route_Test extends TestCase {
+
+	/**
+	 * Holds the instance.
+	 *
+	 * @var Site_Kit_Widget_Permanent_Dismissal_Route
+	 */
+	protected $instance;
+
+	/**
+	 * Sets up the test fixtures.
+	 *
+	 * @return void
+	 */
+	protected function set_up() {
+		parent::set_up();
+
+		Mockery::mock( WP_Error::class );
+
+		$this->instance = new Site_Kit_Widget_Permanent_Dismissal_Route(
+			new Permanently_Dismissed_Site_Kit_Widget_Repository_Fake()
+		);
+	}
+}
