+++ tests/Unit/Dashboard/User_Interface/Constructor_Test.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface;
+
+use Yoast\WP\SEO\Dashboard\Infrastructure\Repositories\Permanently_Dismissed_Site_Kit_Widget_Repository_interface;
+
+/**
+ * Test class for the constructor.
+ *
+ * @group site_kit_widget_permanent_dismissal_route
+ *
+ * @covers Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Widget_Permanent_Dismissal_Route::__construct
+ *
+ * @phpcs:disable Yoast.NamingConventions.ObjectNameDepth.MaxExceeded
+ */
+final class Constructor_Test extends Abstract_Site_Kit_Widget_Permanent_Dismissal_Route_Test {
+
+	/**
+	 * Tests if the needed attributes are set correctly.
+	 *
+	 * @return void
+	 */
+	public function test_construct() {
+		$this->assertInstanceOf(
+			Permanently_Dismissed_Site_Kit_Widget_Repository_interface::class,
+			$this->getPropertyValue( $this->instance, 'permanently_dismissed_site_kit_widget_repository' )
+		);
+	}
+}
