+++ tests/Unit/Dashboard/User_Interface/Get_Conditionals_Test.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface;
+
+use Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Widget_Permanent_Dismissal_Route;
+
+/**
+ * Test class for the get_conditionals method.
+ *
+ * @group site_kit_widget_permanent_dismissal_route
+ *
+ * @covers Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Widget_Permanent_Dismissal_Route::get_conditionals
+ *
+ * @phpcs:disable Yoast.NamingConventions.ObjectNameDepth.MaxExceeded
+ */
+final class Get_Conditionals_Test extends Abstract_Site_Kit_Widget_Permanent_Dismissal_Route_Test {
+
+	/**
+	 * Tests the get_conditionals function.
+	 *
+	 * @covers ::get_conditionals
+	 *
+	 * @return void
+	 */
+	public function test_get_conditionals() {
+		$this->assertEquals( [], Site_Kit_Widget_Permanent_Dismissal_Route::get_conditionals() );
+	}
+}
