+++ tests/Unit/Dashboard/User_Interface/Register_Routes_Test.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface;
+
+use Brain\Monkey\Functions;
+
+/**
+ * Test class for the register_routes method.
+ *
+ * @group site_kit_widget_permanent_dismissal_route
+ *
+ * @covers Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Widget_Permanent_Dismissal_Route::register_routes
+ *
+ * @phpcs:disable Yoast.NamingConventions.ObjectNameDepth.MaxExceeded
+ */
+final class Register_Routes_Test extends Abstract_Site_Kit_Widget_Permanent_Dismissal_Route_Test {
+
+	/**
+	 * Tests the registration of the routes.
+	 *
+	 * @return void
+	 */
+	public function test_register_routes() {
+		Functions\expect( 'register_rest_route' )
+			->once()
+			->with(
+				'yoast/v1',
+				'/site_kit_widget_permanent_dismissal',
+				[
+					[
+						'methods'             => 'POST',
+						'callback'            => [ $this->instance, 'set_site_kit_widget_permanent_dismissal' ],
+						'permission_callback' => [ $this->instance, 'check_capabilities' ],
+						'args'                => [
+							'is_dismissed' => [
+								'required'          => true,
+								'type'              => 'bool',
+								'default'           => false,
+								'sanitize_callback' => 'rest_sanitize_boolean',
+							],
+						],
+					],
+				]
+			);
+
+		$this->instance->register_routes();
+	}
+}
